# postman/postman-collection-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/orders": {
            "post": {
                "description": "Create a new order with order items. Stock quantities will be automatically updated. Authenticated users can only create orders for themselves.",
                "operationId": "create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrderDto" } } }
                },
                "responses": {
                    "201": {
                        "description": "Order successfully created",
                        "content": {
                            "application/json": {
                                "example": {
                                    "order_id": 1,
                                    "user_id": 1,
                                    "order_date": "2025-07-03T03:14:15.000Z",
                                    "status": "pending",
                                    "total_amount": 2599.98,
                                    "shipping_address": "123 Main St, Toronto, ON M1M 1M1",
                                    "payment_method": "Credit Card",
                                    "tracking_number": null,
                                    "notes": "Please deliver after 6 PM",
                                    "created_at": "2025-07-03T03:14:15.000Z",
                                    "updated_at": "2025-07-03T03:14:15.000Z",
                                    "user": { "user_id": 1, "email": "john.doe@example.com", "first_name": "John", "last_name": "Doe" },
                                    "order_items": [
                                        {
                                            "order_item_id": 1,
                                            "product_id": 1,
                                            "quantity": 2,
                                            "unit_price": 1299.99,
                                            "total_price": 2599.98,
                                            "product": { "product_id": 1, "name": "iPhone 15 Pro", "price": 1299.99 }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid input data or insufficient stock" },
                    "401": { "description": "Unauthorized - Invalid or missing token" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Create a new order",
                "tags": ["Orders"]
            },
            "get": {
                "description": "Retrieve orders with pagination, filtering, sorting, searching, and includes. Users see only their own orders unless they are admin/moderator.",
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": { "minimum": 1, "default": 1, "example": 1, "type": "number" }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": { "minimum": 1, "maximum": 100, "default": 10, "example": 10, "type": "number" }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "Number of records to skip (overrides page calculation)",
                        "schema": { "minimum": 0, "example": 0, "type": "number" }
                    },
                    {
                        "name": "sortBy",
                        "required": false,
                        "in": "query",
                        "description": "Field to sort by",
                        "schema": { "example": "created_at", "type": "string" }
                    },
                    {
                        "name": "sortOrder",
                        "required": false,
                        "in": "query",
                        "description": "Sort order",
                        "schema": { "default": "ASC", "example": "ASC", "type": "string", "enum": ["ASC", "DESC"] }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "description": "Search term to filter results",
                        "schema": { "example": "laptop", "type": "string" }
                    },
                    {
                        "name": "include",
                        "required": false,
                        "in": "query",
                        "description": "Related entities to include in the response (joins). Can be comma-separated string or array",
                        "schema": { "example": ["user", "product"], "type": "array", "items": { "type": "array" } }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": [
                                        {
                                            "order_id": 1,
                                            "user_id": 1,
                                            "order_date": "2025-07-03T03:14:15.000Z",
                                            "status": "pending",
                                            "total_amount": 2599.98,
                                            "shipping_address": "123 Main St, Toronto, ON M1M 1M1",
                                            "payment_method": "Credit Card",
                                            "tracking_number": null,
                                            "notes": "Please deliver after 6 PM",
                                            "created_at": "2025-07-03T03:14:15.000Z",
                                            "updated_at": "2025-07-03T03:14:15.000Z",
                                            "user": { "user_id": 1, "email": "john.doe@example.com", "first_name": "John", "last_name": "Doe" },
                                            "order_items": [
                                                {
                                                    "order_item_id": 1,
                                                    "product_id": 1,
                                                    "quantity": 2,
                                                    "unit_price": 1299.99,
                                                    "total_price": 2599.98,
                                                    "product": {
                                                        "product_id": 1,
                                                        "name": "iPhone 15 Pro",
                                                        "price": 1299.99,
                                                        "category": "Electronics",
                                                        "brand": "Apple"
                                                    }
                                                }
                                            ]
                                        }
                                    ],
                                    "pagination": { "page": 1, "limit": 10, "total": 1, "totalPages": 1, "hasNext": false, "hasPrev": false }
                                }
                            }
                        }
                    },
                    "401": { "description": "Unauthorized - Invalid or missing token" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get orders with advanced querying",
                "tags": ["Orders"]
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Retrieve a specific order. Users can only access their own orders unless they are admin/moderator.",
                "operationId": "findOne",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot access other user's orders" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get order by ID",
                "tags": ["Orders"]
            },
            "patch": {
                "description": "Update order information. Users can only update their own orders unless they are admin/moderator.",
                "operationId": "update",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateOrderDto" } } }
                },
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot update other user's orders" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Update order",
                "tags": ["Orders"]
            },
            "delete": {
                "description": "Delete an order. Only admins can delete orders.",
                "operationId": "remove",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Admin access required" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Delete order (Admin only)",
                "tags": ["Orders"]
            }
        },
        "/orders/user/{userId}": {
            "get": {
                "description": "Retrieve orders for a specific user. Users can only access their own orders unless they are admin/moderator.",
                "operationId": "findByUser",
                "parameters": [{ "name": "userId", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot access other user's orders" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get orders by user",
                "tags": ["Orders"]
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "description": "Update order status. Only admin or moderator can change order status.",
                "operationId": "updateStatus",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Insufficient permissions" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Update order status (Admin/Moderator only)",
                "tags": ["Orders"]
            }
        },
...
        "/reviews/user/{userId}": {
            "get": {
                "description": "Retrieve all reviews created by a given user ID. Users can only access their own reviews unless they are admin/moderator.",
                "operationId": "findByUser",
                "parameters": [
                    { "name": "userId", "required": true, "in": "path", "description": "User ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "User reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "user_id": 1,
                                        "product_id": 1,
                                        "rating": 5,
                                        "title": "Great product!",
                                        "comment": "This product exceeded my expectations.",
                                        "is_verified_purchase": true,
                                        "helpful_votes": 5,
                                        "created_at": "2024-01-01T00:00:00Z",
                                        "updated_at": "2024-01-01T00:00:00Z"
                                    }
                                ]
                            }
                        }
                    },
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Cannot access other user's reviews" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Get reviews by a specific user",
                "tags": ["Reviews"]
            }
        },
...
    },
    "info": {
        "title": "eCommerce API",
        "description": "Comprehensive eCommerce API with advanced querying capabilities including pagination, filtering, sorting, searching, and joins",
        "version": "1.0.0",
        "contact": { "name": "API Support", "url": "https://github.com/yourusername/ecommerce-api", "email": "support@yourcompany.com" },
        "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
    },
...
    "servers": [{ "url": "http://localhost:3151", "description": "Development server" }],
    "components": {
        "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
        "schemas": {
...
            "CreateOrderItemDto": {
                "type": "object",
                "properties": {
                    "product_id": { "type": "number", "description": "ID of the product being ordered", "example": 1 },
                    "quantity": { "type": "number", "description": "Quantity of the product", "example": 2, "minimum": 1 },
                    "unit_price": {
                        "type": "number",
                        "description": "Unit price of the product (up to 2 decimal places)",
                        "example": 99.99,
                        "minimum": 0
                    }
                },
                "required": ["product_id", "quantity", "unit_price"]
            },
            "CreateOrderDto": {
                "type": "object",
                "properties": {
                    "user_id": { "type": "number", "description": "ID of the user placing the order", "example": 1 },
                    "shipping_address": {
                        "type": "string",
                        "description": "Complete shipping address for the order",
                        "example": "123 Main St, Toronto, ON M1M 1M1, Canada"
                    },
                    "payment_method": {
                        "type": "string",
                        "description": "Payment method used for the order",
                        "example": "Credit Card",
                        "enum": ["Credit Card", "Debit Card", "PayPal", "Cash", "Bank Transfer"]
                    },
                    "notes": {
                        "type": "string",
                        "description": "Additional notes or instructions for the order",
                        "example": "Please deliver after 6 PM"
                    },
                    "order_items": {
                        "description": "Array of items being ordered",
                        "example": [
                            { "product_id": 1, "quantity": 2, "unit_price": 1299.99 },
                            { "product_id": 2, "quantity": 1, "unit_price": 999.99 }
                        ],
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/CreateOrderItemDto" }
                    }
                },
                "required": ["user_id", "shipping_address", "payment_method", "order_items"]
            },
            "UpdateOrderDto": { "type": "object", "properties": {} },
...
        }
    }
}